cleanup
 - require bundler everywhere
 - stop messing with load paths
 - use Bundler.root
 - do not define helper methods globally
 - remove unnecessary requires
 - ship less files in the gem
