--- spec/support/silence_stream.rb
@@ -1,8 +1,0 @@
-def silence_stream(stream)
-  old_stream = stream.dup
-  stream.reopen('/dev/null')
-  stream.sync = true
-  yield
-ensure
-  stream.reopen(old_stream)
-end
