--- spec/support/system.rb
@@ -1,4 +1,0 @@
-def system!(cmd)
-  rc = system(cmd)
-  fail "failed with exit code #{$?.exitstatus}" if rc.nil? || !rc || $?.exitstatus != 0
-end
